requirements.txt
pynacl
requests
click
cryptography
